refactor: 파일명 컴포넌트 이름과 맞게 수정 및 jsx 제거
