+++ src/pages/Login/LoginPage.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 
-import Input from '../../components/Input';
+import Input from '../../components/Input/Input';
 
 import apis from '../../apis';
