+++ src/pages/Main/Banner/Banner.js
